refactor: extract shared sanitizeNextPath utility and useConnectionsCallback hook
- Move duplicated sanitizeNextPath into @/lib/auth/sanitize-redirect and
  import from youtube/start, youtube/callback, and steam/callback routes
- Extract duplicated connections useEffect into useConnectionsCallback hook
  and replace inline effects in profile-settings-modal and profile-settings-page
- Eliminates variable shadowing of `status` state in profile-settings-page
