+++ apps/web/app/api/users/connections/steam/callback/route.ts
@@ -75,10 +75,7 @@ async function fetchSteamProfile(steamId: string): Promise<{ displayName: string
   return result
 }
 
-function sanitizeNextPath(raw: string): string {
-  if (!raw || !/^\/(?!\/)/.test(raw) || raw.includes("://")) return "/"
-  return raw
-}
+import { sanitizeNextPath } from "@/lib/auth/sanitize-redirect"
 
 function buildRedirect(base: URL, nextPath: string, status: string) {
   const safe = sanitizeNextPath(nextPath)
